fix: solve and simplify the context aware mechanism
