+++ src/corral/agents/prompts/index.json
@@ -22,12 +22,12 @@
   },
   "d880c4d3-fe60-4cf4-813b-2008076cd595": {
     "uuid": "d880c4d3-fe60-4cf4-813b-2008076cd595",
-    "content": "Task Guide: {{task_guide}}\n\n{{examples}}\n\nThink about what to do next and respond in the following format:\n\nThought: <thought>[your reasoning] </thought>\nAction: <action>[tool name]</action>\nAction Input: <action_input>[tool arguments as JSON]</action_input>.\n\nIf you have the final answer, respond with:\nThought: <thought>[your reasoning]</thought>\nFinal Answer: <final_answer>[answer]</final_answer>\n\nThe code expects the XML tags to be used exactly as shown, such that the regex parsing will work.",
+    "content": "Task Guide: {{task_guide}}\n\n{{examples}}\n\nThink about what to do next and respond in the following format:\n\nThought: <thought>[your reasoning]</thought>\nAction: <action>[tool name]</action>\nAction Input: <action_input>[tool arguments as JSON]</action_input>.\n\nIf you have the final answer, respond with:\nThought: <thought>[your reasoning]</thought>\nFinal Answer: <final_answer>[answer]</final_answer>\n\nThe code expects the XML tags to be used exactly as shown. If this format is not followed, the regex parsing will not work.",
     "description": "Very basic REACT prompt. Intended for use with LLM agents.",
     "version": 1,
     "versions": [
       {
-        "content": "Task Guide: {{task_guide}}\n\n{{examples}}\n\nThink about what to do next and respond in the following format:\n\nThought: <thought>[your reasoning] </thought>\nAction: <action>[tool name]</action>\nAction Input: <action_input>[tool arguments as JSON]</action_input>.\n\nIf you have the final answer, respond with:\nThought: <thought>[your reasoning]</thought>\nFinal Answer: <final_answer>[answer]</final_answer>\n\nThe code expects the XML tags to be used exactly as shown, such that the regex parsing will work.",
+        "content": "Task Guide: {{task_guide}}\n\n{{examples}}\n\nThink about what to do next and respond in the following format:\n\nThought: <thought>[your reasoning]</thought>\nAction: <action>[tool name]</action>\nAction Input: <action_input>[tool arguments as JSON]</action_input>.\n\nIf you have the final answer, respond with:\nThought: <thought>[your reasoning]</thought>\nFinal Answer: <final_answer>[answer]</final_answer>\n\nThe code expects the XML tags to be used exactly as shown. If this format is not followed, the regex parsing will not work.",
         "description": "Very basic REACT prompt. Intended for use with LLM agents.",
         "version": 1,
         "created_at": "2025-03-03T12:16:29.129829"
